fix load model problem
